requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
python-dotenv
google-generativeai
requests
plotly